Move proxy handling into UploadFile itself
Per review feedback: rather than threading a proxy-aware client in
from the SQLite caller, have upload.UploadFile resolve the proxy via
config.GetAPIProxy() internally. Drops the UploadFileWithClient seam
and the sqlite-side helper — same fix, smaller surface area.
